+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next
 
+- Change: When the main character is removed from an account, another character is now automatically made the main.
 - Added `/app/v1/player-with-characters/{characterId}` app API endpoint which returns the player account to which 
   the character ID belongs with all characters.
 - Added rate limiting middleware for the application API.